refactor: modularize relational batch executor
