+++ packages/tools/tests/data/relational/query/batch-executor.test.ts
@@ -1,0 +1,8 @@
+/**
+ * Public batch-executor test entrypoint.
+ */
+
+import './batch-executor/chunking.test.js';
+import './batch-executor/retry.test.js';
+import './batch-executor/failure.test.js';
+import './batch-executor/benchmark.test.js';
