+++ packages/tools/tests/data/relational/query/batch-executor/shared.ts
@@ -1,0 +1,6 @@
+import {
+  benchmarkBatchExecution,
+  executeBatchedTask,
+} from '../../../../../src/data/relational/query/batch-executor.js';
+
+export { benchmarkBatchExecution, executeBatchedTask };
